Fixed issues with employees and customer management
Okay so we can now see the lists of employees/customers, also now we can delete customers that dont have any orders pending( im not sure if we should change this functionality or not but it doesnt seem logical to be able to delete customers that have pending orders).
Also for employees, we can now add  employees and can only delete regular employees and not managers, because then we need to cascade the whole branch.(maybe we can fix it? )